yugatool: rename cdc_producer_check to xcluster_consumer_check
The xcluster_consumer_check is intended to run on on a CDC consumer
cluster, and it is specifically meant to check if xcluster replication
is healthy. The new name will hopefully be less confusing.
